refator: 중복 코드 수정 및 refresh 추가
