+++ src/components/PayList.tsx
@@ -15,7 +15,6 @@ interface PayDayProps {
 interface PayListProps {
   startDate: string;
   endDate: string;
-  validDates: string[];
 }
 
 // 아이콘 가져오기 ( {<IconFood/>} 이런식으로 반환됨)
@@ -55,11 +54,7 @@ const PayDay: React.FC<PayDayProps> = ({ data, onClick }) => {
 };
 
 // 전체 소비리스트
-const PayList: React.FC<PayListProps> = ({
-  startDate,
-  endDate,
-  validDates,
-}) => {
+const PayList: React.FC<PayListProps> = ({ startDate, endDate }) => {
   const [loading, setLoading] = useState<boolean>(true); // 로딩 상태 관리
   const { moveToPage } = useMovePage();
   const { spendingRecords, setSpendingData } = useSpendingInfo();
@@ -79,13 +74,6 @@ const PayList: React.FC<PayListProps> = ({
     const ReadConsumption = async () => {
       try {
         setLoading(true);
-        // const hasValidDate = validDates.some(
-        //   (date) => date >= startDate && date <= endDate,
-        // );
-        // if (!hasValidDate) {
-        //   console.warn("소비 기록이 없는 기간이므로 API 호출을 하지 않습니다.");
-        //   return; // API 호출 중단
-        //
 
         if (!startDate || !endDate) {
           if (refresh) {
@@ -105,7 +93,7 @@ const PayList: React.FC<PayListProps> = ({
           return;
         }
         const res = await api.get(
-          `consumption?category="식비"&startDate=${effectiveStartDate}&endDate=${effectiveEndDate}`,
+          `consumption?category=${selectName}&startDate=${effectiveStartDate}&endDate=${effectiveEndDate}`,
         );
 
         console.log(res.data);
@@ -117,7 +105,7 @@ const PayList: React.FC<PayListProps> = ({
         setLoading(false);
       }
     };
-    //ReadConsumption();
+    ReadConsumption();
   }, [startDate, endDate, refresh]);
 
   return (
@@ -146,7 +134,7 @@ const PayList: React.FC<PayListProps> = ({
           </div>
         ))
       ) : (
-        <p className="p-2 text-center text-gray-500">소비 기록이 없습니다.</p>
+        <p className="py-4 text-center text-gray-500">소비 기록이 없습니다.</p>
       )}
     </div>
   );
