--- src/spendingData.ts
@@ -1,49 +1,0 @@
-export const spendingData = {
-  totalPrice: 45000,
-  records: [
-    {
-      year: "2025",
-      month: "2",
-      day: "23",
-      dailyTotal: 25000,
-      details: [
-        {
-          id: 1,
-          price: 12000,
-          category: "식비",
-          details: "점심 - 김밥",
-          point_name: "김밥천국 강남점",
-        },
-        {
-          id: 2,
-          price: 13000,
-          category: "식비",
-          details: "커피",
-          point_name: "스타벅스 홍대입구점",
-        },
-      ],
-    },
-    {
-      year: "2025",
-      month: "2",
-      day: "24",
-      dailyTotal: 20000,
-      details: [
-        {
-          id: 3,
-          price: 8000,
-          category: "교통",
-          details: "지하철",
-          point_name: "서울 2호선 강남역",
-        },
-        {
-          id: 4,
-          price: 12000,
-          category: "식비",
-          details: "저녁 - 라멘",
-          point_name: "이치란 홍대점",
-        },
-      ],
-    },
-  ],
-};
